# models/user.go
package models

import (
	"gorm.io/gorm"
)

type User struct {
	gorm.Model
	Email    string `gorm:"unique;not null" json:"email"`
	Password string `json:"-"`
	IsAdmin  bool   `gorm:"default:false" json:"is_admin"`
}